Migration datatype fix

+++ database/migrations/2022_10_28_153231_create_votes_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('votes', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned();
-            $table->integer('answer_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('answer_id')->unsigned();
             $table->timestamps();
